# documentazione/swagger-openapi.yaml
openapi: 3.0.1
info:
  title: OSP API
  description: Gestionale Occupazione Suolo Pubblico Bari
  version: '1.0'
servers:
  - url: http://localhost:8080/api/osp
    description: Generated server url
paths:
...
  /pratica-management/pratica:
    get:
      tags:
        - pratica-controller
      operationId: getPratiche
      parameters:
        - name: idsMunicipi
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int32
        - name: idsStatiPratica
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int32
        - name: nome
          in: query
          required: false
          schema:
            type: string
        - name: cognome
          in: query
          required: false
          schema:
            type: string
        - name: denominazioneRagSoc
          in: query
          required: false
          schema:
            type: string
        - name: codFiscalePIva
          in: query
          required: false
          schema:
            type: string
        - name: numProtocollo
          in: query
          required: false
          schema:
            type: string
        - name: numProvvedimento
          in: query
          required: false
          schema:
            type: string
        - name: richiestaVerificaRipristinoLuoghi
          in: query
          required: false
          schema:
            type: boolean
        - name: tipologiaProcesso
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: indirizzo
          in: query
          required: false
          schema:
            type: string
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagePraticaDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
    put:
      tags:
        - pratica-controller
      operationId: editPratica
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PraticaInsertEditRequest'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
    post:
      tags:
        - pratica-controller
      operationId: insertPratica
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PraticaInsertEditRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PraticaDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
...
  /pratica-management/date-occupazione:
    put:
      tags:
        - pratica-controller
      operationId: editDateOccupazione
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DateOccupazioneEditRequest'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /egov-management/pratica/aggiornamento-pagamento:
    put:
      tags:
        - egov-controller
      operationId: aggiornamentoPagamento
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PraticaEgovEditRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PraticaDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /egov-management/date-occupazione:
    put:
      tags:
        - egov-controller
      operationId: editDateOccupazione_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DateOccupazioneEgovEditRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PraticaDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
...
  /integrazione-management/rettifica-date:
    post:
      tags:
        - integrazione-controller
      operationId: insertRettificaDate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrazioneDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/IntegrazioneDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
...
  /egov-management/rettifica-date:
    post:
      tags:
        - egov-controller
      operationId: insertRettificaDate_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RettificaDateEgovInsertRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/IntegrazioneDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /egov-management/pratica/validazione-pratica:
    post:
      tags:
        - egov-controller
      operationId: validazionePratica
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PraticaEgovValidazioneRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /egov-management/pratica/rinuncia/inserimento-completo:
    post:
      tags:
        - egov-controller
      operationId: insertRinunciaPraticaCompleta
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RinunciaEgovInsertRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PraticaDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /egov-management/pratica/proroga/inserimento-completo:
    post:
      tags:
        - egov-controller
      operationId: insertProrogaPraticaCompleta
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProrogaEgovInsertRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PraticaDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /egov-management/pratica/inserimento-completo:
    post:
      tags:
        - egov-controller
      operationId: insertPraticaCompleta
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PraticaEgovInsertRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PraticaDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /egov-management/integrazione/inserimento-completo:
    post:
      tags:
        - egov-controller
      operationId: insertIntegrazioneCompleta
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrazioneCompletaEgovInsertRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/IntegrazioneDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
...
  /pratica-management/pratica/{id}:
    get:
      tags:
        - pratica-controller
      operationId: getPratica
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PraticaDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
    delete:
      tags:
        - pratica-controller
      operationId: deletePratica
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
...
  /egov-management/pratica:
    get:
      tags:
        - egov-controller
      operationId: getPratiche_1
      parameters:
        - name: codiceFiscale
          in: query
          required: true
          schema:
            type: string
        - name: codiceProtocollo
          in: query
          required: false
          schema:
            type: string
        - name: idPratica
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PraticaDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /egov-management/pratica/{idPratica}/determina:
    get:
      tags:
        - egov-controller
      operationId: getDetermina
      parameters:
        - name: idPratica
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DeterminaDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /egov-management/pratica/rettificabili:
    get:
      tags:
        - egov-controller
      operationId: getPraticheRettificabili
      parameters:
        - name: codiceFiscalePartitaIva
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /egov-management/pratica/rettifica-date:
    get:
      tags:
        - egov-controller
      operationId: getPraticheRettificaDate
      parameters:
        - name: codiceFiscalePiva
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /egov-management/pratica/avviabili-post-concessione:
    get:
      tags:
        - egov-controller
      operationId: getPraticheAvviabiliPostConcessione
      parameters:
        - name: codiceFiscalePartitaIva
          in: query
          required: true
          schema:
            type: string
        - name: idTipoProcesso
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /egov-management/pratica/attesa-pagamento:
    get:
      tags:
        - egov-controller
      operationId: getPraticheAttesaPagamentoBollo
      parameters:
        - name: codiceFiscalePiva
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
...
components:
  schemas:
...
    PraticaInsertEditRequest:
      required:
        - datiRichiesta
        - firmatario
        - idMunicipio
        - idTipoProcesso
      type: object
      properties:
        id:
          type: integer
          format: int64
        idPraticaOriginaria:
          type: integer
          format: int64
        idUtente:
          type: integer
          format: int64
        idMunicipio:
          type: integer
          format: int32
        motivazioneRichiesta:
          type: string
        firmatario:
          $ref: '#/components/schemas/RichiedenteDTO'
        destinatario:
          $ref: '#/components/schemas/RichiedenteDTO'
        datiRichiesta:
          $ref: '#/components/schemas/DatiRichiestaDTO'
        idTipoProcesso:
          type: integer
          format: int32
        nomeCittadinoEgov:
          type: string
        cognomeCittadinoEgov:
          type: string
        cfCittadinoEgov:
          type: string
        originEgov:
          type: boolean
        marcaBolloPratica:
          $ref: '#/components/schemas/MarcaBolloPraticaDTO'
        marcaBolloDetermina:
          $ref: '#/components/schemas/MarcaBolloPraticaDTO'
...
    DateOccupazioneEditRequest:
      required:
        - dataInizioOccupazione
        - dataScadenzaOccupazione
        - idPratica
      type: object
      properties:
        idPratica:
          type: integer
          format: int64
        dataInizioOccupazione:
          type: string
          format: date
        oraInizioOccupazione:
          $ref: '#/components/schemas/LocalTime'
        dataScadenzaOccupazione:
          type: string
          format: date
        oraScadenzaOccupazione:
          $ref: '#/components/schemas/LocalTime'
        nomeCittadinoEgov:
          type: string
        cognomeCittadinoEgov:
          type: string
        cfCittadinoEgov:
          type: string
        originEgov:
          type: boolean
...
    PraticaEgovEditRequest:
      required:
        - allegati
        - pratica
      type: object
      properties:
        pratica:
          $ref: '#/components/schemas/PraticaEgovInsertEditRequest'
        allegati:
          type: array
          items:
            $ref: '#/components/schemas/AllegatoDTO'
...
    DateOccupazioneEgovEditRequest:
      required:
        - dataInizioOccupazione
        - dataScadenzaOccupazione
      type: object
      properties:
        idPratica:
          type: integer
          format: int64
        numeroProtocollo:
          type: string
        dataInizioOccupazione:
          type: string
          format: date
        oraInizioOccupazione:
          $ref: '#/components/schemas/LocalTime'
        dataScadenzaOccupazione:
          type: string
          format: date
        oraScadenzaOccupazione:
          $ref: '#/components/schemas/LocalTime'
        nomeCittadinoEgov:
          type: string
        cognomeCittadinoEgov:
          type: string
        cfCittadinoEgov:
          type: string
        originEgov:
          type: boolean
...
    PraticaEgovValidazioneRequest:
      type: object
      properties:
        pratica:
          $ref: '#/components/schemas/PraticaInsertEditRequest'
        dimensioneTotaleAllegatiProtocolloKb:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    PagePraticaDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/PraticaDTO'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
...
    DeterminaDTO:
      type: object
      properties:
        mimeType:
          type: string
        nomeFile:
          type: string
        contenuto:
          type: string
...
  securitySchemes:
    jwt:
      type: http
      in: header
      scheme: Bearer